fix(Context): replace uint64_t with uint32_t

+++ src/main/code/context/Context.hpp
@@ -7,7 +7,7 @@
 
 class Context : public Interface {
 public:
-  virtual std::uint64_t executionPoint() const = 0;
+  virtual std::uint32_t executionPoint() const = 0;
   virtual p<StackFrame> stack() const = 0;
   virtual p<JavaValues> locals() const = 0;
   virtual p<JavaValues> arguments() const = 0;
